Gardening: move ThermostatSetpoint into its own header
